TMAG5273: add wake-up & sleep mode
"Wake-up & sleep" is a special mode of the sensor, where the sensor wakes
up in defined intervals ("sleeptime"), performs a readout and generates
an interrupt if the measured value is outside the threshold condition.

After the interrupt is generated, the device is automatically in an
"active" mode, and the last measured value can be read out.

Signed-off-by: Juliane Schulze <juliane.schulze@deveritec.com>