TMAG5273: fix a initialization error after power cycle
There appears to be a bootup issue on the TMAG5273: sometimes, if external
power supply is used, it may be that the sensor won't respond via i2c.

If the power supply is then toggled again, everything works fine.

This behaviour could only be observed on power cycles, every kind of "warm"
restart does not show this behavior. It is independent of the sensor sleep
mode.